Revert "BufferQueue: Increase max slots from 32 to 64"
This reverts commit febd4f4f462444bfcb3f0618d07ac77e3fc1f6ad.

+++ libs/gui/BufferQueue.cpp
@@ -165,7 +165,7 @@ status_t BufferQueue::consumerDisconnect() {
     return mConsumer->disconnect();
 }
 
-status_t BufferQueue::getReleasedBuffers(uint64_t* slotMask) {
+status_t BufferQueue::getReleasedBuffers(uint32_t* slotMask) {
     return mConsumer->getReleasedBuffers(slotMask);
 }
